inserted extra fake axes to appear like x360
There was an issue mention in #1 with games not picking up on the
controller.  We think this is because its not similar enough to an xbox
controller or possibly just that it didn't have a single analog axis
just the DHAT.  Edited the descriptor to include analog stick style axes
and then forced them to always be center because they are not used.

This fixes 1 completely