Fixed issue with boards.txt being outdated
Boards.txt has become extremely outdated because of changes to the Arduino workspace.  This caused issues with compiling this project correctly under newer versions of teensyduino and the arduino environment.  All I did was add the correct menu item to the new boards.txt file from the newer versions.

This closes #5